refactor(vdb): move `ValueType` & `Index` to `types.ts`
