+++ libs/vdb/src/lib/tree/index.ts
@@ -1,2 +1,2 @@
-export * from './leaf-node';
 export * from './internal-node';
+export * from './leaf-node';
