+++ libs/vdb/src/lib/types.ts
@@ -1,0 +1,2 @@
+export type ValueType = boolean | number | string;
+export type Index = number;
